feat: implement create room and get list of rooms
